requirements.txt
streamlit>=1.32.0
openai>=1.3.7
pandas>=1.5.0
numpy>=1.23.0
python-dotenv>=1.0.0
